production: Use iio data_format for ADC/DAC ref
Previously, the full scale values for the ADCs and DACs of tested
boards were hardcoded in `dma_tests.py`, e.g.:

```
def sfdr_low(...):
	...
	ref = 2 ** 12
	...
```

These values are not clearly documented and also differ from board to
board, requiring the tests to be changed manually for each prodtest
image, without any record of these changes anywhere (at most hinted to
by the bash history that remains on the images).

This commit adds `rx_ref` and `tx_ref` properties to the base `rx` and
`tx` classes, respectively, to facilitate the computation of the
reference value based on the IIO channel's reported `data_format`.

A downside is that many channels don't specify a `data_format`, case in
which I chose some hopefully sane defaults, ASSUMING:
- All ADCs have at most 16 bits of resolution
- All DACs have at least 12 bits of resolution

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>