Avoid using genalyzer until it's stable
The genalyzer python bindings currently have multiple incompatible
versions (main vs tfcollins-devel branch). This commit reverts the
dependency on the tfcollins-devel branch to temporarily avoid issues in
installing and using the proper variant.

For genalyzer to be reintroduced, the setup_env script in the
production-tests repo should properly install it. TODO!

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>