Pass harmonic_vals with few specified peaks
In many cases, len(peaks) < len(low), i.e. the test imposes values for more harmonic peaks than are actually detectable above the noise floor, case for which the function was implemented. In the case in which fewer peaks' values are specified than are measured, the test throws an ArrayIndexoutOfBoundsException.

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>